before checkout function
