--- src/components/stripe/CartEmpty.tsx
@@ -1,4 +1,0 @@
-const CartEmpty = () => {
-  return <div>CartEmpty</div>;
-};
-export default CartEmpty;
