+++ src/pages/member/cart.tsx
@@ -1,0 +1,6 @@
+import ViewCart from "@/components/stripe/Cart";
+
+const cart = () => {
+  return <ViewCart />;
+};
+export default cart;
